Finish raw loading animation implementation
