+++ src/components/Button/Button.story.js
@@ -88,7 +88,7 @@ storiesOf(`${GROUPS.COMPONENTS}|Button`, module)
               setTimeout(() => resolve(), 5000);
             })
           }
-          delayMs={1000}
+          delayMs={500}
           primary
         >
           Click me
