+++ src/components/Button/components/LoadingIcon.spec.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import LoadingIcon from './LoadingIcon';
+
+describe.skip('LoadingIcon', () => {
+  it('should ', () => {});
+});
