+++ src/components/Button/components/index.js
@@ -1,0 +1,2 @@
+export { default as Button } from './Button';
+export { default as LoadingIcon } from './LoadingIcon';
